+++ internal/logic/balance/balance.go
@@ -8,26 +8,26 @@ import (
 	"unibee/utility"
 )
 
-func UserBalanceDetailQuery(ctx context.Context, merchantId uint64, userId int64, gatewayId int64) (*ro.GatewayUserDetailQueryInternalResp, error) {
+func UserBalanceDetailQuery(ctx context.Context, merchantId uint64, userId int64, gatewayId uint64) (*ro.GatewayUserDetailQueryInternalResp, error) {
 	user := query.GetUserAccountById(ctx, uint64(userId))
 	merchant := query.GetMerchantInfoById(ctx, merchantId)
-	gateway := query.GetGatewayById(ctx, gatewayId)
+	gateway := query.GetGatewayById(ctx, uint64(gatewayId))
 	utility.Assert(user != nil, "user not found")
 	utility.Assert(merchant != nil, "merchant not found")
 
-	queryResult, err := api.GetGatewayServiceProvider(ctx, int64(gateway.Id)).GatewayUserDetailQuery(ctx, gateway, userId)
+	queryResult, err := api.GetGatewayServiceProvider(ctx, gateway.Id).GatewayUserDetailQuery(ctx, gateway, userId)
 	if err != nil {
 		return nil, err
 	}
 	return queryResult, nil
 }
 
-func MerchantBalanceDetailQuery(ctx context.Context, merchantId uint64, gatewayId int64) (*ro.GatewayMerchantBalanceQueryInternalResp, error) {
+func MerchantBalanceDetailQuery(ctx context.Context, merchantId uint64, gatewayId uint64) (*ro.GatewayMerchantBalanceQueryInternalResp, error) {
 	merchant := query.GetMerchantInfoById(ctx, merchantId)
-	gateway := query.GetGatewayById(ctx, gatewayId) // todo mark 根据 MerchantId 配置 Gateway
+	gateway := query.GetGatewayById(ctx, uint64(gatewayId))
 	utility.Assert(merchant != nil, "merchant not found")
 
-	queryResult, err := api.GetGatewayServiceProvider(ctx, int64(gateway.Id)).GatewayMerchantBalancesQuery(ctx, gateway)
+	queryResult, err := api.GetGatewayServiceProvider(ctx, gateway.Id).GatewayMerchantBalancesQuery(ctx, gateway)
 	if err != nil {
 		return nil, err
 	}
